add changes to routes and schemas of alerts and thresholds
