+++ src/bemserver_api/resources/alerts/schemas.py
@@ -2,7 +2,7 @@ import marshmallow as ma
 import marshmallow_sqlalchemy as msa
 
 from bemserver_api import AutoSchema
-from bemserver_core.model.alerts import Alert , AlertType, Threshold
+from bemserver_core.model.alerts import Alert , AlertType
 
 class AlertSchema(AutoSchema):
     class Meta(AutoSchema.Meta):
@@ -12,7 +12,7 @@ class AlertSchema(AutoSchema):
     alert_type = ma.fields.Enum(AlertType)
     threshold_id = msa.auto_field()
     actual_consumption = msa.auto_field()
-    timestamp = msa.auto_field(dump_only=True)
+    timestamp = msa.auto_field()
     device_id = msa.auto_field()
     user_id = msa.auto_field()
 
@@ -23,12 +23,3 @@ class AlertQueryArgsSchema(ma.Schema):
     threshold = ma.fields.Float()
     actual_consumption = ma.fields.Float()
 
-class ThresholdSchema(AutoSchema):
-    class Meta(AutoSchema.Meta):
-        model = Threshold
-
-    id = msa.auto_field(dump_only=True)
-    value = msa.auto_field(required=True)
-    device_id = msa.auto_field(required=True)
-    user_id = msa.auto_field(required=True)
-    created_at = msa.auto_field(dump_only=True)
