+++ src/bemserver_api/resources/thresholds/__init__.py
@@ -1,0 +1,4 @@
+from .routes import blp  # noqa
+
+def register_blueprints(api):
+    api.register_blueprint(blp)
